Refactor #118 (#121)
Some suggested changes for PR #118
