--- belgif-rest-problem-it/belgif-rest-problem-openapi-validation-spring-boot-2-it/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-server.servlet.context-path=/spring
-
-spring.jackson.default-property-inclusion=non_null
-
-#If this property is not set the client will receive an InvalidRequestException or InvalidResponseException
-#without knowing what is wrong with the request / response as the message field will be missing.
-# https://bitbucket.org/atlassian/swagger-request-validator/src/master/swagger-request-validator-spring-webmvc/README.md#Configuration
-server.error.include-message=always
-
-spring.mvc.throw-exception-if-no-handler-found=true
