--- belgif-rest-problem-it/belgif-rest-problem-openapi-validation-spring-boot-2-it/src/test/resources/logback-test.xml
@@ -1,32 +1,0 @@
-<configuration>
-
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%d [%thread] %.-1level %logger{36} %M - %msg%n</pattern>
-    </encoder>
-    <filter class="ch.qos.logback.classic.filter.LevelFilter">
-      <level>ERROR</level>
-      <onMatch>DENY</onMatch>
-      <onMismatch>ACCEPT</onMismatch>
-    </filter>
-    <target>System.out</target>
-  </appender>
-
-  <appender name="STDERR" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%d [%thread] %.-1level %logger{36} %M - %msg%n</pattern>
-    </encoder>
-    <filter class="ch.qos.logback.classic.filter.LevelFilter">
-      <level>ERROR</level>
-      <onMatch>ACCEPT</onMatch>
-      <onMismatch>DENY</onMismatch>
-    </filter>
-    <target>System.err</target>
-  </appender>
-
-  <root level="info">
-    <appender-ref ref="STDOUT" />
-    <appender-ref ref="STDERR" />
-  </root>
-
-</configuration>
