--- belgif-rest-problem-it/belgif-rest-problem-openapi-validation-spring-boot-3-it/.gitignore
@@ -1,1 +1,0 @@
-**/openapi.yaml
