--- belgif-rest-problem-it/belgif-rest-problem-openapi-validation-spring-boot-3-it/src/main/java/io/github/belgif/rest/problem/Application.java
@@ -1,12 +1,0 @@
-package io.github.belgif.rest.problem;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
