+++ belgif-rest-problem-it/belgif-rest-problem-spring-boot-2-it/src/test/java/io/github/belgif/rest/problem/RestProblemOpenApiValidationSpringBoot2IT.java
@@ -8,18 +8,19 @@ import io.restassured.RestAssured;
 import io.restassured.specification.RequestSpecification;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class RestProblemOpenapiValidationSpringBoot2IT extends AbstractOpenApiValidationSpringBootIT {
+class RestProblemOpenApiValidationSpringBoot2IT extends AbstractOpenApiValidationSpringBootIT {
 
     @LocalServerPort
     private int port;
 
     @Override
     protected RequestSpecification getSpec() {
-        return RestAssured.with().baseUri("http://localhost").port(port).basePath("/spring");
+        return RestAssured.with().baseUri("http://localhost").port(port).basePath("/spring/openapi-validation");
     }
 
     @Override
-    @Disabled("Only seems to work in SB3. In SB2 the atlassian lib throws the same exception, but the missing path seems not to be handled by a controller and not caught by the ControllerAdvice.")
+    @Disabled("Only seems to work in SB3. In SB2 the atlassian lib throws the same exception, " +
+            "but the missing path seems not to be handled by a controller and not caught by the ControllerAdvice.")
     void notFound() {
         super.notFound();
     }
