+++ belgif-rest-problem-it/belgif-rest-problem-spring-boot-3-it/src/main/java/io/github/belgif/rest/problem/OpenApiValidationController.java
@@ -1,15 +1,17 @@
 package io.github.belgif.rest.problem;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.github.belgif.rest.problem.openapi.validation.sb2.api.MyFirstPathApi;
-import io.github.belgif.rest.problem.openapi.validation.sb2.api.MyHeaderPathApi;
-import io.github.belgif.rest.problem.openapi.validation.sb2.api.MyQueryPathApi;
-import io.github.belgif.rest.problem.openapi.validation.sb2.model.MyRequestBodySchema;
+import io.github.belgif.rest.problem.openapi.validation.sb3.api.MyFirstPathApi;
+import io.github.belgif.rest.problem.openapi.validation.sb3.api.MyHeaderPathApi;
+import io.github.belgif.rest.problem.openapi.validation.sb3.api.MyQueryPathApi;
+import io.github.belgif.rest.problem.openapi.validation.sb3.model.MyRequestBodySchema;
 
 @RestController
-public class ValidationController implements MyFirstPathApi, MyHeaderPathApi, MyQueryPathApi {
+@RequestMapping("/openapi-validation")
+public class OpenApiValidationController implements MyFirstPathApi, MyHeaderPathApi, MyQueryPathApi {
 
     @Override
     public ResponseEntity<String> myFirstGetOperation(String myParam) {
