+++ belgif-rest-problem-it/belgif-rest-problem-spring-boot-3-it/src/main/resources/application.properties
@@ -1,2 +1,7 @@
 server.servlet.context-path=/spring
 io.github.belgif.rest.problem.scan-additional-problem-packages=com.acme.custom
+
+#If this property is not set the client will receive an InvalidRequestException or InvalidResponseException
+#without knowing what is wrong with the request / response as the message field will be missing.
+# https://bitbucket.org/atlassian/swagger-request-validator/src/master/swagger-request-validator-spring-webmvc/README.md#Configuration
+server.error.include-message=always
