+++ jacoco-aggregator/pom.xml
@@ -73,18 +73,6 @@
     </dependency>
     <dependency>
       <groupId>io.github.belgif.rest.problem</groupId>
-      <artifactId>belgif-rest-problem-openapi-validation-spring-boot-2-it</artifactId>
-      <version>${project.version}</version>
-      <type>war</type>
-    </dependency>
-    <dependency>
-      <groupId>io.github.belgif.rest.problem</groupId>
-      <artifactId>belgif-rest-problem-openapi-validation-spring-boot-3-it</artifactId>
-      <version>${project.version}</version>
-      <type>war</type>
-    </dependency>
-    <dependency>
-      <groupId>io.github.belgif.rest.problem</groupId>
       <artifactId>belgif-rest-problem-jackson-minimal-it</artifactId>
       <version>${project.version}</version>
     </dependency>
